requirements.txt
# ==================================================
# DEPENDENCIAS PARA SISTEMA DE GESTIÓN DE TAREAS
# Sistema con Microservicios + JWT + MySQL
# ==================================================

# Framework web principal
Flask==2.3.3
Werkzeug==2.3.7

# Base de datos MySQL
mysql-connector-python==8.1.0
PyMySQL==1.1.0

# Autenticación JWT
PyJWT==2.8.0
cryptography==41.0.4

# Hashing de contraseñas
bcrypt==4.0.1

# Requests para comunicación entre microservicios
requests==2.31.0
urllib3==2.0.4

# Variables de entorno
python-dotenv==1.0.0

# Utilidades de fecha y hora
python-dateutil==2.8.2

# Validación de datos
marshmallow==3.20.1

# CORS para API (opcional)
Flask-CORS==4.0.0

# Logging avanzado (opcional)
colorlog==6.7.0

# Testing (opcional)
pytest==7.4.2
pytest-flask==1.2.0

# Documentación API (opcional)
flask-restx==1.1.0

# ==================================================
# INSTALACIÓN
# ==================================================
# pip install -r requirements.txt
#
# O instalar individualmente:
# pip install Flask mysql-connector-python PyJWT bcrypt requests python-dotenv
#
# ==================================================
# VERSIONES MÍNIMAS REQUERIDAS
# ==================================================
# Python >= 3.8
# MySQL >= 8.0 (recomendado) o >= 5.7
# 
# ==================================================
# OPCIONAL - PARA DESARROLLO
# ==================================================
# autopep8==2.0.4      # Formateo de código
# flake8==6.1.0        # Linting
# black==23.9.1        # Formateo de código
# isort==5.12.0        # Ordenar imports